refactor(cli): extract workload profile to dedicated module
Rollback trigger: test failure regressions on cargo or clippy in workload module
